-added ant targets for generating tags/ebrowse info and for executing arbitrary jam targets -added -O1 to DEV compilation flags -checkpoint--utils and constraint engine module tests passing
git-svn-id: https://europa-pso.googlecode.com/svn/PLASMA/branches/enum-branch@5160 23e1bc29-d954-0410-9766-47de34158ba0

+++ src/PLASMA/Resource/component/NDDL/InterpreterResources.cc
@@ -21,9 +21,9 @@ namespace EUROPA {
   {
     Id<NDDL::NddlResource>  instance = (new NDDL::NddlResource(planDb, objectType, objectName,true))->getId();
 
-    std::vector<float> argValues;
+    std::vector<double> argValues;
     for (unsigned int i=0;i<arguments.size();i++)
-      argValues.push_back((float)(arguments[i]->getSingletonValue()));
+      argValues.push_back(arguments[i]->getSingletonValue());
 
     if (argValues.size() == 0)
       instance->constructor();
@@ -74,9 +74,9 @@ namespace EUROPA {
   {
     Id<NDDL::NddlReusable>  instance = (new NDDL::NddlReusable(planDb, objectType, objectName,true))->getId();
 
-    std::vector<float> argValues;
+    std::vector<double> argValues;
     for (unsigned int i=0;i<arguments.size();i++)
-      argValues.push_back((float)(arguments[i]->getSingletonValue()));
+      argValues.push_back(arguments[i]->getSingletonValue());
 
     if (argValues.size() == 0)
       instance->constructor();
@@ -128,9 +128,9 @@ namespace EUROPA {
   {
     Id<NDDL::NddlReservoir>  instance = (new NDDL::NddlReservoir(planDb, objectType, objectName,true))->getId();
 
-    std::vector<float> argValues;
+    std::vector<double> argValues;
     for (unsigned int i=0;i<arguments.size();i++)
-      argValues.push_back((float)(arguments[i]->getSingletonValue()));
+      argValues.push_back(arguments[i]->getSingletonValue());
 
     if (argValues.size() == 0)
       instance->constructor();
@@ -194,9 +194,9 @@ namespace EUROPA {
   {
     Id<NDDL::NddlUnary>  instance = (new NDDL::NddlUnary(planDb, objectType, objectName,true))->getId();
 
-    std::vector<float> argValues;
+    std::vector<double> argValues;
     for (unsigned int i=0;i<arguments.size();i++)
-      argValues.push_back((float)(arguments[i]->getSingletonValue()));
+      argValues.push_back(arguments[i]->getSingletonValue());
 
     if (argValues.size() == 0)
       instance->constructor();
