+++ src/PLASMA/Resource/component/NDDL/NddlResource.hh
@@ -45,7 +45,7 @@ namespace NDDL {
 
     virtual void close();
 
-    virtual void constructor(float c_max);
+    virtual void constructor(double c_max);
     virtual void constructor();
 
     void handleDefaults(bool autoClose = true);
@@ -85,9 +85,9 @@ namespace NDDL {
 
     virtual void close();
 
-    virtual void constructor(float c, float ll_min);
-    virtual void constructor(float c, float ll_min, float cr_max);
-    virtual void constructor(float c, float ll_min, float c_max, float cr_max);
+    virtual void constructor(double c, double ll_min);
+    virtual void constructor(double c, double ll_min, double cr_max);
+    virtual void constructor(double c, double ll_min, double c_max, double cr_max);
     virtual void constructor();
 
     void handleDefaults(bool autoClose = true);
@@ -132,11 +132,11 @@ namespace NDDL {
 
     virtual void close();
     
-    virtual void constructor(float ic, float ll_min, float ll_max);
+    virtual void constructor(double ic, double ll_min, double ll_max);
     
-    virtual void constructor(float ic, float ll_min, float ll_max, float p_max, float c_max);
+    virtual void constructor(double ic, double ll_min, double ll_max, double p_max, double c_max);
     
-    virtual void constructor(float ic, float ll_min, float ll_max, float pr_max, float p_max, float cr_max, float c_max);
+    virtual void constructor(double ic, double ll_min, double ll_max, double pr_max, double p_max, double cr_max, double c_max);
     
     virtual void constructor();
 
@@ -216,11 +216,11 @@ namespace NDDL {
 
     virtual void close();
     
-    virtual void constructor(float ic, float ll_min, float ll_max);
+    virtual void constructor(double ic, double ll_min, double ll_max);
     
-    virtual void constructor(float ic, float ll_min, float ll_max, float p_max, float c_max);
+    virtual void constructor(double ic, double ll_min, double ll_max, double p_max, double c_max);
     
-    virtual void constructor(float ic, float ll_min, float ll_max, float pr_max, float p_max, float cr_max, float c_max);
+    virtual void constructor(double ic, double ll_min, double ll_max, double pr_max, double p_max, double cr_max, double c_max);
     
     virtual void constructor();
